+++ CHANGELOG.md
@@ -10,6 +10,7 @@
 * Add setting to enable spellcheck in 'code' scope
 
 ### Fixed
+* Fix the command line becoming too long when there are hundreds of source roots in the module
 * Fix graphic insertion wizard not opening after pasting an image from the clipboard
 * Fix exception #3967
 * Make sure the external packagge inclusion cache refills are not attempted unnecessarily